Move mofa handling to chats mod
