+++ src/data/store.rs
@@ -8,17 +8,11 @@ use super::search::SortCriteria;
 use super::{chats::Chats, downloads::Downloads, search::Search};
 use anyhow::Result;
 use chrono::{DateTime, Utc};
-use makepad_widgets::{error, Action, ActionDefaultRef, Cx, DefaultNone};
+use makepad_widgets::{Action, ActionDefaultRef, DefaultNone};
 use moly_backend::Backend;
-use moly_mofa::{AgentId, MofaServerId};
-use moly_mofa::{
-    MofaAgent,
-    MofaClient, TestServerResponse,
-};
+
 use moly_protocol::data::{Author, DownloadedFile, File, FileID, Model, ModelID, PendingDownload};
-use std::collections::HashMap;
 use std::rc::Rc;
-use std::sync::mpsc;
 
 pub const DEFAULT_MAX_DOWNLOAD_THREADS: usize = 3;
 const DEFAULT_MOFA_ADDRESS: &str = "http://localhost:8000";
@@ -28,13 +22,6 @@ pub enum StoreAction {
     Search(String),
     ResetSearch,
     Sort(SortCriteria),
-    None,
-}
-
-#[derive(Debug, DefaultNone, Clone)]
-pub enum MoFaTestServerAction {
-    Success(String),
-    Failure(Option<String>),
     None,
 }
 
@@ -68,27 +55,7 @@ pub struct Store {
     pub search: Search,
     pub downloads: Downloads,
     pub chats: Chats,
-    pub preferences: Preferences,
-    pub mofa_servers: HashMap<MofaServerId, ServerInfo>,
-    pub available_agents: HashMap<AgentId, MofaAgent>,
-}
-
-#[derive(Clone, Debug)]
-pub struct ServerInfo {
-    pub address: String,
-    pub client: Rc<MofaClient>,
-    pub connection_status: MofaServerConnectionStatus,
-    // TODO(Julian): remove this
-    pub available_agent_ids: Vec<AgentId>,
-}
-
-/// The connection status of the server
-#[derive(Debug, Clone, PartialEq)]
-pub enum MofaServerConnectionStatus {
-    Connecting,
-    Connected,
-    Disconnected,
-    // Disconnecting,
+    pub preferences: Preferences
 }
 
 impl Default for Store {
@@ -118,8 +85,6 @@ impl Store {
             downloads: Downloads::new(backend.clone()),
             chats: Chats::new(backend),
             preferences,
-            mofa_servers: HashMap::new(),
-            available_agents: HashMap::new(),
         };
 
         store.downloads.load_downloaded_files();
@@ -131,7 +96,7 @@ impl Store {
         store.search.load_featured_models();
 
         if moly_mofa::should_be_real() && moly_mofa::should_be_visible() {
-            store.register_mofa_server(DEFAULT_MOFA_ADDRESS.to_string());
+            store.chats.register_mofa_server(DEFAULT_MOFA_ADDRESS.to_string());
         }
         store
     }
@@ -198,8 +163,8 @@ impl Store {
             match entity_id {
                 ChatEntityId::Agent(agent_id) => {
                     if let (Some(client), Some(agent)) = (
-                        self.get_client_for_agent(agent_id),
-                        self.available_agents.get(agent_id)
+                        self.chats.get_client_for_agent(agent_id),
+                        self.chats.available_agents.get(agent_id)
                     ) {
                         chat.send_message_to_agent(agent, prompt, &client);
                     } else {
@@ -218,51 +183,6 @@ impl Store {
                 }
             }
         }
-    }
-
-    pub fn agents_list(&mut self) -> Vec<MofaAgent> {
-        // Here we should ask each client for its available agents
-        // and return the union of all agents
-        // TODO(Julian): remove cloning and server_id should be server_address, set from within the client
-        if self.available_agents.is_empty() {
-            println!("no agents fetched, fetching from servers");
-            for server in self.mofa_servers.values() {
-                let server_agents = server.client.get_available_agents();
-                println!("extending agents list with server {} agents: {}", server.address, server_agents.len());
-                // agents.extend(server_agents);
-                for mut agent in server_agents {
-                    let unique_agent_id = unique_agent_id(&agent.id, &server.address);
-                    agent.server_id = moly_mofa::MofaServerId(server.address.clone());
-                    agent.id = unique_agent_id.clone();
-                    self.available_agents.insert(unique_agent_id, agent);
-                }
-            }
-        }
-        // TODO(Julian): remove unnecessary cloning, rework this
-        self.available_agents.values().cloned().collect()
-    }
-
-    /// Tests the connection to a MoFa server by requesting /v1/models
-    /// The connection status is updated at the App level based on the actions dispatched
-    pub fn test_mofa_server_connection(&mut self, address: String) {
-        self.mofa_servers.get_mut(&MofaServerId(address.to_string())).unwrap().connection_status = MofaServerConnectionStatus::Connecting;
-        let (tx, rx) = mpsc::channel();
-        if let Some(server) = self.mofa_servers.get(&MofaServerId(address.to_string())) {
-            server.client.test_connection(tx.clone());
-        }
-
-        std::thread::spawn(move || match rx.recv() {
-            Ok(TestServerResponse::Success(server_address)) => {
-                Cx::post_action(MoFaTestServerAction::Success(server_address));
-            }
-            Ok(TestServerResponse::Failure(server_address)) => {
-                Cx::post_action(MoFaTestServerAction::Failure(Some(server_address)));
-            }
-            Err(e) => {
-                error!("Error receiving response from MoFa backend: {:?}", e);
-                Cx::post_action(MoFaTestServerAction::Failure(None));
-            }
-        });
     }
 
     pub fn edit_chat_message(&mut self, message_id: usize, updated_message: String) {
@@ -304,7 +224,7 @@ impl Store {
                 .find(|df| df.file.id == *file_id)
                 .map(|df| Some(df.file.name.clone()))?,
             Some(ChatEntityId::Agent(agent)) => {
-                self.available_agents
+                self.chats.available_agents
                     .get(&agent)
                     .map(|a| a.name.clone())
             }
@@ -438,49 +358,4 @@ impl Store {
         // support having no chat selected
         self.init_current_chat();
     }
-
-    pub fn remove_mofa_server(&mut self, address: &str) {
-        self.mofa_servers.remove(&MofaServerId(address.to_string()));
-        self.available_agents.retain(|_, agent| agent.server_id.0 != address);
-    }
-
-    pub fn register_mofa_server(&mut self, address: String) -> MofaServerId {
-        let server_id = MofaServerId(address.clone());
-        let client = Rc::new(MofaClient::new(address.clone()));
-        
-        self.mofa_servers.insert(server_id.clone(), ServerInfo {
-            address: address.clone(),
-            client,
-            available_agent_ids: Vec::new(),
-            connection_status: MofaServerConnectionStatus::Disconnected,
-        });
-
-        // TODO(Julian): we might want to these in one step
-        self.test_mofa_server_connection(address);
-        self.fetch_agents_from_server(&server_id);
-        
-        server_id
-    }
-
-    pub fn fetch_agents_from_server(&mut self, server_id: &MofaServerId) {
-        // TODO(Julian): remove cloning and server_id should be server_address, set from within the client
-        let server = self.mofa_servers.get(server_id).unwrap();
-        let agents = server.client.get_available_agents();
-        for mut agent in agents {
-            let unique_agent_id = unique_agent_id(&agent.id, &server.address);
-            agent.server_id = moly_mofa::MofaServerId(server.address.clone());
-            agent.id = unique_agent_id.clone();
-            self.available_agents.insert(unique_agent_id, agent);
-        }
-    }
-
-    pub fn get_client_for_agent(&self, agent_id: &AgentId) -> Option<Rc<MofaClient>> {
-        self.available_agents.get(agent_id)
-            .and_then(|agent| self.mofa_servers.get(&agent.server_id))
-            .map(|server| server.client.clone())
-    }
-}
-
-fn unique_agent_id(agent_id: &AgentId, server_address: &str) -> AgentId {
-    AgentId(format!("{}-{}", agent_id.0, server_address))
 }
